CERSComparison prep for new problem types.

+++ Examples/SimpleCERESComparison/src/CeresSolver.h
@@ -2,7 +2,7 @@
 #include <vector>
 class CeresSolver {
 public:
-	CeresSolver(int N)
+	CeresSolver(size_t N)
 	{
         functionData.resize(N);
 	}
